Force exit if errors are encountered while sending discovery packets
In some rare circumstances, UDP send operations can fail:

```
    thread 'main' panicked at src/discover/mod.rs:251:45:
    called `Result::unwrap()` on an `Err` value: SendError { .. }
    note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

    WARN netlink_proto::connection: failed to forward message to connection handle: channel closed
```

Which leaves the process without a discovery task and not replying to UDP messages anymore.

This commit forces the process to exit so it can be restarted by systemd.

Systemd service configuration is also being adjusted so `Restart` is now `always`, restarting the process even if the exit code is `0`.