Resolve spurious freezes due mutual blocking
DynamicTimer and the main Server task could deadlock in situations where UDP Advertise packets arrived too quickly
and two calls to `batch_timer.set_timeout` were performed before the DynamicTimer task was able to run.

In this scenario, the main Server task would be blocked and unable to consume the next expiration, blocking the
DynamicTimer from progressing as well.

This change modifies the DynamicTimer task so it:

a) Always consumes the next timeout while waiting for the previosly configured one.

This has the advantage of updating the batch timer if multiple events and generating new timeout requests.

b) Never block while attempting to send a timeout notification.

If the previously timeout communication was not consumed yet, print a log message and wait for more timeout requests.